+++ FirebaseVertexAI/CHANGELOG.md
@@ -22,6 +22,9 @@
   is now optional (`Int?`); it may be `null` in cases such as when a
   `GenerateContentRequest` contains only images or other non-text content.
   (#13721)
+- [changed] The default request timeout is now 180 seconds instead of the
+  platform-default value for a `URLRequest`; this timeout may still be
+  customized in `RequestOptions`.
 
 # 11.3.0
 - [added] Added `Decodable` conformance for `FunctionResponse`. (#13606)